Calculate thumbnail size based on screen width
Using the wrong thumbnail size will result in ugly black space
on the right side, or bad spacing between items, so a static
size is unacceptable. Instead, aim for approximately 160px per
item, but adjust that based on the width to minimize the unused
space on each row.

Thumbnails are always loaded at the maximum size (160x160), but
scaled based on the current width, so they will be appropriately
sized regardless of orientation, and never reloaded or re-cached.

Fixes NEMO#189 and NEMO#212