Fix N#123 - Errant images displayed in transition to ImagePage
When viewing three sequential images in the image page, going
back to the list, and loading the second or third of those images
again, the previous image will be shown on top of the list during
the transition to ImagePage before eventually being replaced with
the actual image.

The fix is simply to enable clipping on the image view.