Fix image jittering during pinch-to-zoom
When zooming, the contentWidth/Height of the flickable are changed,
which also requires changing contentX and contentY to preserve
position. However, doing this in onPinchUpdated causes it to happen
before other bindings and values have updated, resulting in the
image visibly shaking on screen during the zoom.

Rather than updating them explicitly from onPinchUpdated, bind the
flickable contentWidth and contentHeight to the scaled image size,
and update the contentX and contentY when those change. This means
everything (scale, contentWidth/Height, and contentX/Y) changes in
the same frame, and there is no jittering.