Fix and document incorrect usage of ImagePage visibleIndex
Despite the name, this property is never changed and only determines
the index shown when first loading the page. It's a bit complex to
change without breaking something, so use middle.index instead to
refer to the current index for now. Documented this so we can fix
it later.

This fixes a slideshow bug causing it to always start from the image
the page was loaded for, not the current image.