NEMO#260 Gallery can now show and play videos with basic control
First of all, a video player with basic controls has been added (VideoPlayer.qml).
The video player doesn't have any control due to a flickering issue which appear when any UI component
in the same view as the video player is drawn. GalleryModel now uses File as rootType and it has a filter
on video/* and image/* mimeTypes to be able to show both videos and images. Moreover, a QGLWidget has
been set as QDeclarativeView's viewport to avoid lag issues which would occur for the whole lifecycle of
the application after playing the first video.
Two different mouse areas are now used to detect clicks on the images when the image is zoomed in and
when it is not. That is to avoid conflict between the many mouse areas used in this application.
Many more small adaptation changes have also been applied. Hacky solutions have been documented in the
code using comments.