Fixes a bug with the images not being correctly aligned in one case
This fix makes image align properly.
Previously, images were not aligned correctly when the videoplayer was
started in portrait mode, and was popped after rotating the phone in
landscape mode. This was due to flickTo NumberAnimation breaking the binding
on leftMost.x.

A new Binding element has been added to control this behaviour, and
a new bool property (keepMiddleItemCentered) has been added to control that
binding. The developer has to set keepMiddleItemCentered to false when he
wants to modify the x value of the leftMost item manually, and he has to set
it back to true once he has finished such modifications, to be sure that the
leftMost item is correctly positioned, thus showing the "middle" item in the
center of the screen.