Acquire video playback resources with libresourceqt
Missing resources may be responsible for intermittent lag seen with
xvideo (where videos play at 1-2fps without any apparent reason).
This is an attempt to help with that problem. Regardless, requesting
the policy is good practice.

Due to the need for a QObject to manage libresourceqt, main.cpp was
also refactored into a Gallery class at the same time.