NEMO#402 Exiting slideshow mode doesn't change the index of the displayed element
I've done some changes to the slideshow mode code, cleaning it a bit
and refactoring where was needed (visibleIndex wasn't really representing
the index of the displayed element).

Also I've refactored ImagePage code (here too, visibleIndex wasn't really
representing the index of the displayed element).

Moreover, I've changed the slideshow code to make it so that when
you pop the slideshow page, you won't be taken to the element where
you started the slideshow mode from.
This means that if you now exit slideshow mode when element with index i is on screen,
you'll see element with index i in the list view mode too, as asked in the bug thread
of NEMO#402