Restructured/simplified ImagePage's and ImageContainer's MouseAreas
The code to handle taps on the image has been moved from ImagePage
to ImageContainer.
The same has been done for the code which shows the video player.

Some logic has been rewritten, and useless code deleted.

Also, a toolbarTimer has been added. The timer is needed to have
the possibility to handle double taps without triggering the code
which shows/hide the toolbars (statusBar and toolBar)