ImageContainer restructured to be gallery model independent
ImageContainer code has been changed so that it doesn't require
an existing gallery model anymore. This change is needed to make
ImageContainer reusable for use cases where a gallery model is not
available (e.g. when showing an image which is not in one of the
directories the media tracker tracks)