Hack -> image is still while fullscreen mode is toggled
I really didn't like the fact that the image was resized when
it was zoomed in and the fullscreen mode was toggled
(i.e. toolbar and statusBar appear while the image is zoomed in).
This is because the Page component is resized when those bars
appear.
The changes I made are hacks that make the image appear
still and centered on the screen, and make the toolbar
act as overlaid items, which are shown on top of the image.

This results in a MUCH better experience, in my humble opinion.