NEMO#196 new feature: gallery handles external images/videos
Gallery's components have been modified to make the gallery the
default handler of images/videos. This means the gallery will be
started whenever a 3rd party app calls Qt.openUrlExternally(imageUrl)
or similar.

A new component (SingleImagePage) has been added to handle that
use case, taking into account that the media element the user wants
us display could be in a folder which is NOT tracked by the Tracker
(i.e. we can't use a DocumentGalleryItem to retrieve info about
the media element)